feat: обобщен контейнер для кеша зависимостей формул
