+++ docs/en/docs/evaluate_rules.md
@@ -38,4 +38,4 @@
 
 ## Ссылки на связанные файлы:
 
-- [Кеш контейнер для хранения зависимостей](../server/apps/dcis/helpers/sheet_cache.py)
+- [Кеш контейнер для хранения зависимостей](../server/apps/dcis/helpers/sheet_formula_cache.py)
